feat: add routes to app
