+++ src/global.css
@@ -40,55 +40,3 @@ a {
   border-left: 1px solid #ebeef8;
   border-right: 1px solid #ebeef8;
 }
-
-.new-tweet-form {
-  padding: 1.5rem 1.25rem;
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.new-tweet-form label {
-  display: flex;
-  gap: 0.75rem;
-}
-
-.new-tweet-form label img {
-  width: 3rem;
-  height: 3rem;
-  border-radius: 9999px;
-}
-
-.new-tweet-form label textarea {
-  flex: 1;
-  border: 0;
-  font-size: 1.25rem;
-  font-weight: 500;
-  margin-top: 0.75rem;
-  resize: none;
-}
-
-.new-tweet-form label textarea:focus {
-  outline: none;
-}
-
-.new-tweet-form label textarea::placeholder {
-  color: #5b7083;
-}
-
-.new-tweet-form button[type="submit"] {
-  background: var(--twitter-blue);
-  color: #FFF;
-  font-size: 1rem;
-  font-weight: 900;
-  margin-left: auto;
-  padding: 0.75rem 1.5rem;
-  border-radius: 9999px;
-  border: 0;
-  transition: all 0.3s ease-in;
-}
-
-.new-tweet-form button[type="submit"]:hover {
-  filter: brightness(0.9);
-}
-
